added internationalization support to kwm2. I haven't increased the patch level, since Matthias did this very recently.
Note: it looks like, this is the first time, that an application needs
the bNLS flag to writeEntry. It does not work for me. The kwmrc get's
longer and longer with every run.

Just for information:
We can not write Desktop7= entries, since we can not decide later, if this
is the default or the national entry. So KConfig writes Desktop7[C]=, if
no language is set.