MD: Almost done with work on kcmdisplay.
Color schemes substantially rewritten. Schemes are now installed in separate
kcsrc files. Users can add, remove and alter their personal schemes. The
current scheme is given a separate listing in the scheme list so quick,
simple changes are easy to make. ( This also ensures long term compatibility
with root window property storage for V1.0+ )

Changed font list creation. Two modes for recovery of font lists are
available - general and fixed. Both use the KDE font list if it is available
and cross check with the fonts available on the server.