Clear out the passwords in the password files. We don't need them in kgreeter.cpp, and we don't want the encrypted passwords to be stored in the core file in case of a SEGV.
I'm not so sure if setting the passord entry to zero is a good thing, cause
that's a blank passwd -> you don't have to enter a passord. On the other
hand this code should _not_ verify the password itself.

Anyway.. Maybe clearing the passord entry and setting it to "*" would be
better? Comments?