A small fix for 1.0 compatibility (just two strcmp replaced with stricmp)
Reason : Some users still have device kdelnks with
FSType=default in it. The check was only for FSType=Default.

In fact, at some point (well, between 1.46 and 1.47 exactly),
"Mount default" was replaced with
tmp.sprintf( "%s Default", klocale->translate( "Mount" ) );
so we now have to handle both !

Guess how long it takes to find out that the log message "Torben: Monster
bug fix commit :-) " was the one that changed this :) [I'm used to it now !]