[IEM] Fixing bug #1096
Original bug info

"Package: kde
Version: 1.1 (Debian binaries built 1999-03-06)

By default, the "hover" pop-up message for the "box of tools" icon (yellow
box, plus hammer and wrench) for /usr/share/applnk/Utilities/.directory, is
"No comment available".  This is not likely to get the people in suits to
throw Windows out of our office (which is, after all, the idea).  It
probably needs to be something more dynamic (preferably with words like
"pro-active" and "multidisciplinary", to really get the suits excited :-) )."