Theme manager now handles konsole "theming".  The format is roughly:
[Konsole]
schema=<SCHEMA_FILE>
background=<WALLPAPER>

The wallpaper must be the wallpaper specified in the schema file.  I
couldn't figure out a way to easily modify the schema file as it is
not KConfig controlled (why?).

I bumped the version number 1.0.1.  If that's not a good idea, feel
free to bump it back down.